Add alarm time and date variables
Alarm menu now show time and date. Next estep is edit those values pressing menu button to change between hour, day, etc... Then with up and down buttons the values que be changed and enter to enter a submenu asking if the alarm should be on or off, then if selects on the alarm will be saved, if chooses off, the alarm will return to the current hour or last alarm set by user.